Modify 11689

+++ Baekjoon Online Judge/11689.cpp
@@ -13,10 +13,10 @@ int main() {
     ll n;
     cin >> n;
     ll ans=n;
-    for(int i=2; i*i<=n; i++) {
+    for(ll i=2; i*i<=n; i++) {
         if(n%i==0) {
             ans -= ans/i;
-            while(n%i==0) n/= i;
+            while(n%i==0) n /= i;
         }
     }
     if(n>1) ans -= ans/n;
